+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [v2.1.0] - 02.02.2022
+
+### Added
+
+- `TimePicker`-field for ACF
+- `Checkbox`-field for ACF
+
 ## [v2.0.1] - 26.02.2022
 
 ## Fixed
@@ -48,6 +55,7 @@
 
 - from bitbucket to github ( going opensource :muscle: :heart_eyes: )
 
+[v2.1.0]: https://github.com/towa-digital/towa-acf-fields/releases/tag/v2.1.0
 [v2.0.1]: https://github.com/towa-digital/towa-acf-fields/releases/tag/v2.0.1
 [v2.0.0]: https://github.com/towa-digital/towa-acf-fields/releases/tag/v2.0.0
 [v1.1.1]: https://github.com/towa-digital/towa-acf-fields/releases/tag/v1.1.1